examples/mcp/streamable-http/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
    "semantic-kernel>=1.32.2",
]
